+++ Integrations/RSANetWitness_v11_1/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 - Fixed an issue with ***fetch-incidents*** where setting a *Fetch Limit* would drop older incidents if the number of the fetched incidents was greater than the limit.
-- Added new argument to *pageNumber* to ***netwitness-get-incidents*** command. The new argument allows the user to get incidents from a specific page, intended to be used with limit argument.
+- Added the new argument *pageNumber* to  the ***netwitness-get-incidents*** command. The new argument allows the user to get incidents from a specific page and is intended to be used with limit argument.
 
 ## [19.12.1] - 2019-12-25
   - Fixed an issue where environment proxy effects the integration when even though no proxy should be used.